final? pipeline

+++ scripts/luigi/multi-day/recording_processor.py
@@ -17,26 +17,28 @@ from spikeinterface import create_sorting_analyzer
 import spikeinterface.full as si
 import spikeinterface.curation as scur
 
-import torch
+from pprint import pprint
 
 
 N_JOBS = os.cpu_count() // 2
 
-# Example paths for testing
-EXAMPLE_PATHS = {
-    'source_dir':  Path('/mnt/d/temp_processing'),
-    'working_dir': Path('/mnt/d/temp_processing'),
-    # 'source_dir':  Path("/Users/vigji/Desktop/short_recording_oneshank"),
-    # 'working_dir': None,
-    'test_recording': (
-        Path("/Volumes/Extreme SSD/2024-11-13_14-39-11"),
-        "Record Node 111#Neuropix-PXI-110.ProbeA"
-    )
-}
-
 #################################
 # Utils
 #################################
+
+def get_stream_name(folder: Path) -> str:
+    """Get stream name from folder."""
+    STREAM_NAME_MATCH = "Record Node */experiment*/recording*/continuous/Neuropix-PXI-*.Probe*"
+    try:
+        stream_path = next(folder.glob(STREAM_NAME_MATCH))
+    except StopIteration:
+        raise ValueError(f"No stream found in {folder} matching {STREAM_NAME_MATCH}")
+    
+    parts = stream_path.parts
+    record_node = next(part for part in parts if part.startswith("Record Node "))
+    record_node_number = record_node.split("Record Node ")[1]
+    probe = stream_path.name
+    return f"Record Node {record_node_number}#{probe}"
 
 def _n_seconds_to_formatted_time(n_seconds: int) -> str:
     """Convert seconds to formatted time string."""
@@ -47,8 +49,18 @@ def _n_seconds_to_formatted_time(n_seconds: int) -> str:
 
 def is_timestamp_folder(folder: Path) -> bool:
     """Check if a folder name matches the timestamp pattern."""
+    if not folder.is_dir():
+        return False
     timestamp_pattern = re.compile(r"^\d{4}-\d{2}-\d{2}_\d{2}-\d{2}-\d{2}$")
-    return bool(timestamp_pattern.match(folder.name)) and folder.is_dir()
+    if not timestamp_pattern.match(folder.name):
+        return False
+    try:
+        get_stream_name(folder)
+    except ValueError:
+        return False
+    
+    return True
+
 
 def copy_folder(src: str | Path, dst: str | Path, overwrite: bool = True) -> Path:
     """Copy a folder to a new location.
@@ -85,20 +97,6 @@ def copy_folder(src: str | Path, dst: str | Path, overwrite: bool = True) -> Pat
     
     return dst
 
-def get_stream_name(folder: Path) -> str:
-    """Get stream name from folder."""
-    STREAM_NAME_MATCH = "Record Node */experiment*/recording*/continuous/Neuropix-PXI-*.Probe*"
-    try:
-        stream_path = next(folder.glob(STREAM_NAME_MATCH))
-    except StopIteration:
-        raise ValueError(f"No stream found in {folder} matching {STREAM_NAME_MATCH}")
-    
-    parts = stream_path.parts
-    record_node = next(part for part in parts if part.startswith("Record Node "))
-    record_node_number = record_node.split("Record Node ")[1]
-    probe = stream_path.name
-    return f"Record Node {record_node_number}#{probe}"
-
 
 
 ##################################
@@ -109,19 +107,20 @@ def get_stream_name(folder: Path) -> str:
 class RecordingProcessor:
     """Handles path management and processing state for a recording."""
     source_path: Path
-    working_dir: Path  # Optional[Path]
+    root_working_dir: Path | None  # Optional[Path]
     stream_name: str
     is_split_recording: bool = False
-    
-    def __post_init__(self):
-        # self.source_path = Path(self.source_path)
-        # If working_dir is None, use the parent directory of the source_path
-        # self.working_dir = Path(self.working_dir) if self.working_dir is not None else self.source_path.parent
 
-        self.local_folder = self.working_dir / self.source_path.name
+    def __post_init__(self):
+        self.source_path = Path(self.source_path)
+        self.root_working_dir = Path(self.root_working_dir) if self.root_working_dir is not None else None
+        
+        # Initialize all paths
+        self.processed_in_different_folder = self.root_working_dir is not None
+        self.local_folder = self.root_working_dir / self.source_path.name if self.processed_in_different_folder else self.source_path
         self.kilosort_folder = self.local_folder / "kilosort4"
         self.analyzer_folder = self.kilosort_folder / "analyser"
-    
+
     def try_load_recording(self) -> OpenEphysBinaryRecordingExtractor | None:
         """Try to load the recording, return None if it fails."""
         try:
@@ -157,12 +156,16 @@ class RecordingProcessor:
     
     def copy_to_working_dir(self) -> Path:
         """Copy recording data to working directory."""
-        self.local_folder = copy_folder(self.source_path, self.working_dir)
+        if self.processed_in_different_folder:
+            self.local_folder = copy_folder(self.source_path, self.root_working_dir)
         return self.local_folder
 
     def copy_results_to_source_dir(self, overwrite: bool = True) -> None:
         """Copy results to source directory, only copying new or modified files."""
-        copy_folder(self.local_folder, self.source_path.parent, overwrite=overwrite)
+        if self.processed_in_different_folder:
+            copy_folder(self.local_folder, self.source_path.parent, overwrite=overwrite)
+        else:
+            print(f"Would copy results to source directory, but we are in the same folder: {self.local_folder}")
 
     def remove_sorting_results(self) -> None:
         """Remove sorting results."""
@@ -174,7 +177,10 @@ class RecordingProcessor:
 
     def cleanup_working_dir(self) -> None:
         """Cleanup working directory."""
-        shutil.rmtree(self.local_folder)
+        if self.processed_in_different_folder:
+            shutil.rmtree(self.local_folder)
+        else:
+            print(f"Would cleanup working directory, but we are in the same folder: {self.local_folder}")
 
 
 #################################
@@ -203,9 +209,9 @@ def standard_preprocessing(recording_extractor: OpenEphysBinaryRecordingExtracto
 def run_sorting(recording: OpenEphysBinaryRecordingExtractor, kilosort_folder: Path, overwrite: bool) -> si.BaseSorting:
     sorting = ss.run_sorter(
         sorter_name="kilosort4",
-        recording=recording,
+        recording=recording, 
         folder=kilosort_folder,
-        remove_existing_folder=overwrite,
+        remove_existing_folder=True,
         n_jobs=N_JOBS,
         verbose=True,
     )
@@ -226,7 +232,7 @@ def compute_stats(sorting_ks: si.BaseSorting, recording: OpenEphysBinaryRecordin
     analyzer.compute("random_spikes", method="uniform", max_spikes_per_unit=500, **job_kwargs)
     analyzer.compute("waveforms", ms_before=1.0, ms_after=2.0, **job_kwargs)
     analyzer.compute("templates", operators=["average", "median", "std"], **job_kwargs)
-
+    print(job_kwargs)
     si.compute_noise_levels(analyzer, **job_kwargs)
     si.compute_spike_amplitudes(analyzer, **job_kwargs)
 
@@ -247,7 +253,7 @@ def compute_stats(sorting_ks: si.BaseSorting, recording: OpenEphysBinaryRecordin
 
 def process_recording(processor: RecordingProcessor, dry_run: bool = False, overwrite: bool = False, copy_upstream: bool = True, cleanup: bool = True) -> None:
     """Process a single recording."""
-    print(f"\nProcessing {processor.source_path.name} with stream name {processor.stream_name}")
+    print(f"\nProcessing {processor.source_path} with stream name {processor.stream_name}")
     global_t_start = time.time()
     
     # Step 1: Try to load or copy recording data
@@ -337,10 +343,10 @@ def process_recording(processor: RecordingProcessor, dry_run: bool = False, over
 #################################
 
 
-def find_recordings_to_process(source_dir: Path, working_dir: Path) -> list[RecordingProcessor]:
+def find_recordings_to_process(source_dir: Path, root_working_dir: Path) -> list[RecordingProcessor]:
     """Find all recordings to process."""
     source_dir = Path(source_dir)
-    working_dir = source_dir.parent if working_dir is None else Path(working_dir)
+    root_working_dir = Path(root_working_dir) if root_working_dir is not None else None
     processors = []
 
     # Find all timestamp folders up to 3 levels deep using explicit glob patterns for speed
@@ -351,7 +357,8 @@ def find_recordings_to_process(source_dir: Path, working_dir: Path) -> list[Reco
         if is_timestamp_folder(folder)
     )
     
-    print(f"Found {len(timestamp_folders)} timestamp folders to process")
+    print(f"Found {len(timestamp_folders)} timestamp folders to process: ")
+    pprint(timestamp_folders)
     
     for timestamp_folder in timestamp_folders:
         try:
@@ -362,14 +369,14 @@ def find_recordings_to_process(source_dir: Path, working_dir: Path) -> list[Reco
                 if not any(p.source_path == parent_folder for p in processors):
                     processors.append(RecordingProcessor(
                         source_path=parent_folder,
-                        working_dir=working_dir,
+                        root_working_dir=root_working_dir,
                         stream_name=stream_name,
                         is_split_recording=True
                     ))
             else:
                 processors.append(RecordingProcessor(
                     source_path=timestamp_folder,
-                    working_dir=working_dir,
+                    root_working_dir=root_working_dir,
                     stream_name=stream_name
                 ))
         except Exception as e:
@@ -379,26 +386,49 @@ def find_recordings_to_process(source_dir: Path, working_dir: Path) -> list[Reco
     
     return processors
 
+@dataclass
+class RunArgs:
+    source_dir: Path
+    root_working_dir: Path
+    dry_run: bool
+    overwrite: bool
+    copy_upstream: bool
+    cleanup: bool
+
+luigi_macbook_kwargs = RunArgs(
+    source_dir=Path("/Users/vigji/Desktop/short_recording_oneshank"),  # Path("/Volumes/Extreme SSD/mouse_data_electrode_tips"),
+    root_working_dir=Path("/Users/vigji/Desktop/temp_dir"),
+    dry_run=False,
+    overwrite=False,
+    copy_upstream=True,
+    cleanup=True
+)
+
+wsl_run_kwargs = RunArgs(
+    source_dir=Path("/mnt/d/temp_processing"),
+    root_working_dir=Path("/mnt/d/temp_processing"),
+    dry_run=True,
+    overwrite=False,
+    copy_upstream=True,
+    cleanup=True
+)
+
+
 def main():
     """Example usage of the recording processor."""
-    print(EXAMPLE_PATHS)
+    run_kwargs = luigi_macbook_kwargs
     # Example with dry run
     
     # Example with real processing
     print("\n=== Real Processing Example ===")
-    processors = find_recordings_to_process(
-        "/Volumes/SystemsNeuroBiology/SNeuroBiology_shared/P07_PREY_HUNTING_YE/e01_ephys _recordings/M29",
-        working_dir=None,
-        # EXAMPLE_PATHS['source_dir'],
-        # EXAMPLE_PATHS['working_dir']
-    )
-    # print(processors)
+    processors = find_recordings_to_process(run_kwargs.source_dir, run_kwargs.root_working_dir)
+    pprint(processors)
     first_split = next((p for p in processors if p.is_split_recording), None)
     first_not_split = next((p for p in processors if not p.is_split_recording), None)
     
     for processor in [p for p in [first_split, first_not_split] if p is not None]:
         print("\n" + "-"*100 + "\n")
-        process_recording(processor, dry_run=True, overwrite=False)
+        process_recording(processor, dry_run=run_kwargs.dry_run, overwrite=run_kwargs.overwrite, copy_upstream=run_kwargs.copy_upstream, cleanup=run_kwargs.cleanup)
 
 if __name__ == "__main__":
     main()
